requirements.txt
openai
spacy
speechrecognition
streamlit
pyyaml
pandas
plotly
numpy
transformers
torch
sentence-transformers
